Add quit function (Temporarily inserting delay)

+++ include/sdfw.h
@@ -11,7 +11,7 @@ namespace sdfw
 {
     void init();
 
-    void openWindow(uint16_t width, uint16_t height);
+    void openWindow(uint32_t width, uint32_t height);
 
     void quit();
 }
