+++ source/sdfw.cpp
@@ -17,7 +17,7 @@ namespace sdfw
         sdfwEngine::get()->init();
     }
 
-    void openWindow(uint16_t width, uint16_t height)
+    void openWindow(uint32_t width, uint32_t height)
     {
         SDFW_ENGINE(Socket)->execOpenWindow(width, height);
     }
